+++ src/components/LayoutComponents/Sidebar.jsx
@@ -1,8 +1,9 @@
 import React, { useRef, useEffect } from "react";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import "../styles/sidebar.css";
-import useIsMobile from "../hooks/useIsMobile";
+import "../../styles/ForLayout/sidebar.css";
+import useIsMobile from "../../hooks/useIsMobile";
 import { NavLink } from "react-router-dom";
+import ThemeSwitch from "./ThemeSwitch";
 
 const Sidebar = ({ isExpanded, items, toggleSidebar, toggleSidebarRef }) => {
     const sidebarRef = useRef(null);
@@ -24,18 +25,21 @@ const Sidebar = ({ isExpanded, items, toggleSidebar, toggleSidebarRef }) => {
         return () => {
             document.removeEventListener("mousedown", handleClickOutside);
         };
-    }, [isExpanded, toggleSidebar, toggleSidebarRef]);
-
-    return (
+    }, [isExpanded, toggleSidebar, toggleSidebarRef]);    return (
         <div ref={sidebarRef} className={`sidebar ${isExpanded ? 'expanded' : ''}`}>
-            {items.map((item, index) => (
-                <NavLink key={index} to={item.path} className="sidebar-button">
-                    <button key={index} >
-                        <FontAwesomeIcon icon={item.icon} className="sidebar-icon" />
-                        <span className="sidebar-text">{item.text}</span>
-                    </button>
-                </NavLink>
-            ))}
+            <div className="sidebar-content">
+                {items.map((item, index) => (
+                    <NavLink key={index} to={item.path} className="sidebar-button">
+                        <button key={index}>
+                            <FontAwesomeIcon icon={item.icon} className="sidebar-icon" />
+                            <span className="sidebar-text">{item.text}</span>
+                        </button>
+                    </NavLink>
+                ))}
+            </div>
+            <div className="sidebar-footer">
+                <ThemeSwitch />
+            </div>
         </div>
     );
 };
