+++ src/pages/Login.jsx
@@ -1,87 +1,3 @@
-// import React from "react";
-// import { signInWithEmailAndPassword } from "firebase/auth";
-// import { auth } from "../firebase"; // adjust path as necessary
-// import "../styles/login.css";
-// import image from "../assets/pics/login-2.png";
-// import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-// import { faUser, faKey } from '@fortawesome/free-solid-svg-icons';
-// import { useNavigate } from "react-router-dom";
-// import ThemeSwitch from "../components/ThemeSwitch";
-
-
-// function Login(){
- 
-//     const navigate = useNavigate();
-
-//     const handleLogin = async (e) => {
-//       e.preventDefault();
-//       const email = e.target[0].value;
-//       const password = e.target[1].value;
-//       const status = document.querySelector(".status");
-    
-//       try {
-//         const userCredential = await signInWithEmailAndPassword(auth, email, password);
-//         const user = userCredential.user;
-    
-//         // Navigate based on user's email or role logic
-//         if (user.email === "admin123@gmail.com") {
-//           navigate("/admin", {replace: true});
-//         } else if (user.email === "manager123@gmail.com") {
-//           navigate("/manager", {replace: true});
-//         } else {
-//           status.innerHTML = "לא מורשה להיכנס למערכת זו";
-//         }
-//       } catch (error) {
-//         status.innerHTML = "שם משתמש או סיסמה שגויים";
-//         console.error("Login error:", error.message);
-//       }
-//     };
-    
-
-
-//     return (
-//       <div className="container">
-//         {/* Login Form */}
-//         <div className="right-panel">
-//           <div className="top-panel">
-//             <div className="login-box">
-//               <h2>כניסה למערכת</h2>
-//               <form onSubmit={handleLogin}>
-//                 <div className="input-group">
-//                   <FontAwesomeIcon icon={faUser} className="icon" />
-//                   <input type="email" placeholder="דואר אלקטרוני" required />
-//                 </div>
-//                 <div className="input-group">
-//                   <FontAwesomeIcon icon={faKey} className="icon" />
-//                   <input type="password" placeholder="סיסמה" required />
-//                 </div>
-//                 <div className="status"><span></span></div>
-//                 <button className="submit-button" type="submit">כניסה</button>
-//               </form>
-//             </div>
-//           </div>
-//           <div className="bottom-panel">
-//             <ThemeSwitch />
-//           </div>
-//         </div>
-
-//         {/* Title & Image */}
-//         <div className="left-panel">
-//           <div className="title-box">
-//             <h1>מערכת ניהול מלאי</h1>
-//             <h2>ותיקי מטה יהודה</h2>
-//           </div>
-//           <div className="illustration">
-//             <img src={image} alt="מחסן" />
-//           </div>
-//         </div>
-//       </div>
-//     );
-//   }
-  
-//   export default Login;
-
-
 import React, { useState, useEffect } from "react";
 import { signInWithEmailAndPassword } from "firebase/auth";
 import { auth } from "../firebase/firebase"; // Using firebase from firebase folder
@@ -92,7 +8,7 @@ import image from "../assets/pics/login-2.png";
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faUser, faKey, faExclamationTriangle } from '@fortawesome/free-solid-svg-icons';
 import { useNavigate } from "react-router-dom";
-import ThemeSwitch from "../components/ThemeSwitch";
+import ThemeSwitch from "../components/LayoutComponents/ThemeSwitch";
 import { ROLES } from "../constants/roles.js";
 
 function Login() {
