--- src/pages/Signup.jsx
@@ -1,37 +1,0 @@
-// src/pages/Signup.jsx
-import { useState } from "react";
-import { auth } from "../firebase/firebase";
-import { createUserWithEmailAndPassword } from "firebase/auth";
-
-function Signup() {
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-
-  const handleSignup = async (e) => {
-    e.preventDefault();
-    try {
-      await createUserWithEmailAndPassword(auth, email, password);
-      alert("User registered!");
-    } catch (err) {
-      alert(err.message);
-    }
-  };
-
-  return (
-    <form onSubmit={handleSignup}>
-      <input
-        type="email"
-        placeholder="Email"
-        onChange={(e) => setEmail(e.target.value)}
-      />
-      <input
-        type="password"
-        placeholder="Password"
-        onChange={(e) => setPassword(e.target.value)}
-      />
-      <button type="submit">Sign Up</button>
-    </form>
-  );
-}
-
-export default Signup;
