+++ src/styles/ForLayout/footer.css
@@ -13,6 +13,8 @@
   color: #333;
   font-size: 1.05rem;
   box-shadow: 0 -2px 8px rgba(0,0,0,0.03);
+
+  z-index: 400;
 }
 
 .footer-content {
