+++ src/styles/ForModals/productModal.css
@@ -1,4 +1,4 @@
-.addProduct-widget {
+.Product-modal {
   display: flex;
   flex-direction: column;
   width: 90%;
@@ -16,7 +16,7 @@
   to { opacity: 1; transform: translateY(0); }
 }
 
-.addProduct-widget h2 {
+.Product-modal h2 {
   font-size: 1.5rem;
   font-weight: 600;
   margin-bottom: var(--spacing-lg);
@@ -24,28 +24,28 @@
   text-align: center;
 }
 
-.addProduct-form {
+.Product-form {
   display: flex;
   flex-direction: column;
   width: 100%;
   gap: var(--spacing-md);
 }
 
-.addProduct-form-group {
+.Product-form-group {
   display: flex;
   flex-direction: column;
   margin-bottom: var(--spacing-sm);
   gap: var(--spacing-xs);
 }
 
-.addProduct-form-group label {
+.Product-form-group label {
   font-weight: 500;
   color: var(--secondary-text);
   text-align: right;
 }
 
-.addProduct-form-group input,
-.addProduct-form-group select {
+.Product-form-group input,
+.Product-form-group select {
   direction: rtl;
   width: 100%;
   padding: var(--spacing-sm) var(--spacing-md);
@@ -57,20 +57,20 @@
   transition: border-color var(--transition-fast), box-shadow var(--transition-fast);
 }
 
-.addProduct-form-group input:focus,
-.addProduct-form-group select:focus {
+.Product-form-group input:focus,
+.Product-form-group select:focus {
   outline: none;
   border-color: var(--primary);
   box-shadow: 0 0 0 3px rgba(76, 175, 80, 0.25);
 }
 
-.addProduct-form-group input[readonly] {
+.Product-form-group input[readonly] {
   background-color: var(--bg-color);
   cursor: not-allowed;
 }
 
 /* Group container for buttons */
-.addProduct-button-group {
+.Product-button-group {
   display: flex;
   gap: var(--spacing-md);
   margin-top: var(--spacing-lg);
@@ -78,7 +78,7 @@
   width: 100%;
 }
 
-.addNewProduct-button {
+.NewProduct-button {
   border-radius: var(--border-radius-md);
   cursor: pointer;
   padding: var(--spacing-sm) var(--spacing-lg);
@@ -92,62 +92,62 @@
   min-width: 120px;
 }
 
-.addNewProduct-button:first-child {
+.NewProduct-button:first-child {
   background-color: var(--primary);
   color: white;
 }
 
-.addNewProduct-button:first-child:hover {
+.NewProduct-button:first-child:hover {
   background-color: var(--primary-dark);
 }
 
-.addNewProduct-button:last-child {
+.NewProduct-button:last-child {
   background-color: transparent;
   border: 1px solid var(--border-color);
   color: var(--secondary-text);
 }
 
-.addNewProduct-button:last-child:hover {
+.NewProduct-button:last-child:hover {
   background-color: var(--bg-color);
   color: var(--primary-text);
 }
 
 /* Status message */
-.addProduct-widget p {
+.Product-modal p {
   margin-top: var(--spacing-md);
   text-align: center;
   color: var(--primary);
   font-weight: 500;
 }
 
-.addProduct-widget p.error {
+.Product-modal p.error {
   color: var(--danger);
 }
 
 /* Responsive adjustments */
 @media (min-width: 768px) {
-  .addProduct-form-group {
+  .Product-form-group {
     flex-direction: row;
     align-items: center;
   }
   
-  .addProduct-form-group label {
+  .Product-form-group label {
     min-width: 120px;
     margin-bottom: 0;
   }
 }
 
 @media (max-width: 576px) {
-  .addProduct-widget {
+  .Product-modal {
     padding: var(--spacing-md);
     width: 95%;
   }
   
-  .addProduct-button-group {
+  .Product-button-group {
     flex-direction: column;
   }
   
-  .addNewProduct-button {
+  .NewProduct-button {
     width: 100%;
   }
 }
