+++ src/styles/ForWidgets/categoryWidget.css
