+++ src/styles/ForWidgets/profileWidget.css
