+++ src/styles/theme.css
@@ -97,7 +97,8 @@
 body {
   background-color: var(--bg-color);
   color: var(--primary-text);
-  font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
+  font-family: Arial, sans-serif;
+  /* font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif; */
   transition: background-color var(--transition-normal), color var(--transition-normal);
   margin: 0;
   padding: 0;
